@@ -19,83 +18,60 @@
     "test:cov": "jest --coverage",
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
     "test:e2e": "jest --config ./test/jest-e2e.json",
-    "prettify": "prettier --write \"src/**/*.ts\"",
-    "typeorm": "ts-node -r tsconfig-paths/register ./node_modules/typeorm/cli.js --config src/config/typeorm.config.ts",
-    "seed:create": "ts-node -r tsconfig-paths/register ./node_modules/typeorm/cli.js --config src/config/seed.config.ts \"migration:create\" -n",
-    "seed:run": "ts-node -r tsconfig-paths/register ./node_modules/typeorm/cli.js --config src/config/seed.config.ts \"migration:run\"",
-    "seed:revert": "ts-node -r tsconfig-paths/register ./node_modules/typeorm/cli.js --config src/config/seed.config.ts \"migration:revert\"",
-    "migration": "npm run typeorm migration:run",
-    "migration:create": "typeorm migration:create -n",
-    "migration:revert": "npm run typeorm migration:revert"
-  },
-  "prettier": {
-    "singleQuote": false,
-    "tabWidth": 2,
-    "printWidth": 120,
-    "trailingComma": "es5"
+    "typeorm": "typeorm-ts-node-commonjs",
+    "seed:create": "func() { npm run typeorm migration:create \"./src/seed/$1\"; }; func",
+    "seed:run": "typeorm-ts-node-commonjs migration:run -d ./src/config/seed.config.ts",
+    "seed:revert": "typeorm-ts-node-commonjs migration:revert -d ./src/config/seed.config.ts",
+    "migration:run": "typeorm-ts-node-commonjs migration:run -d ./src/config/typeorm.config.ts",
+    "migration:create": "func() { npm run typeorm migration:create \"./src/migrations/$1\"; }; func",
+    "migration:revert": "typeorm-ts-node-commonjs migration:revert -d ./src/config/typeorm.config.ts"
   },
   "dependencies": {
-    "@nestjs-modules/mailer": "^1.6.0",
-    "@nestjs/common": "^7.6.15",
-    "@nestjs/config": "^1.1.7",
-    "@nestjs/core": "^7.6.15",
-    "@nestjs/jwt": "^7.2.0",
-    "@nestjs/mapped-types": "*",
-    "@nestjs/passport": "^7.1.5",
-    "@nestjs/platform-express": "^7.6.15",
-    "@nestjs/serve-static": "^2.1.4",
-    "@nestjs/typeorm": "^7.1.5",
-    "@types/uuid": "^8.3.0",
-    "aws-sdk": "^2.1029.0",
-    "bcrypt": "^5.0.1",
-    "check-video": "^1.0.2",
-    "class-transformer": "^0.4.0",
-    "class-validator": "^0.13.1",
-    "cls-hooked": "^4.2.2",
-    "handlebars": "^4.7.7",
-    "jwt-decode": "^3.1.2",
-    "lodash": "^4.17.21",
-    "moment": "^2.29.1",
-    "multer": "^1.4.2",
-    "multer-s3": "^2.10.0",
-    "node-request-context": "^1.0.5",
-    "nodemailer": "^6.5.0",
-    "passport": "^0.4.1",
-    "passport-jwt": "^4.0.0",
-    "passport-local": "^1.0.0",
-    "pg": "^8.5.1",
+    "@nestjs-modules/mailer": "^1.8.1",
+    "@nestjs/common": "^9.0.0",
+    "@nestjs/config": "^2.3.1",
+    "@nestjs/core": "^9.0.0",
+    "@nestjs/jwt": "^10.0.2",
+    "@nestjs/passport": "^9.0.3",
+    "@nestjs/platform-express": "^9.0.0",
+    "@nestjs/swagger": "^6.2.1",
+    "@nestjs/typeorm": "^9.0.1",
+    "bcrypt": "^5.1.0",
+    "class-transformer": "^0.5.1",
+    "class-validator": "^0.14.0",
+    "moment": "^2.29.4",
+    "passport": "^0.6.0",
+    "passport-jwt": "^4.0.1",
+    "pg": "^8.10.0",
     "reflect-metadata": "^0.1.13",
-    "rimraf": "^3.0.2",
-    "rxjs": "^6.6.6",
-    "typeorm": "^0.2.32",
-    "typeorm-polymorphic": "0.0.10",
-    "uuid": "^8.3.2"
+    "rxjs": "^7.2.0",
+    "typeorm": "^0.3.12",
+    "webpack": "^5.78.0"
   },
   "devDependencies": {
-    "@nestjs/cli": "^7.6.0",
-    "@nestjs/schematics": "^7.3.0",
-    "@nestjs/testing": "^7.6.15",
-    "@types/express": "^4.17.11",
-    "@types/jest": "^26.0.22",
-    "@types/lodash": "^4.14.168",
-    "@types/multer": "^1.4.5",
-    "@types/node": "^14.14.36",
-    "@types/passport-jwt": "^3.0.5",
-    "@types/passport-local": "^1.0.33",
-    "@types/supertest": "^2.0.10",
-    "@typescript-eslint/eslint-plugin": "^4.19.0",
-    "@typescript-eslint/parser": "^4.19.0",
-    "eslint": "^7.22.0",
-    "eslint-config-prettier": "^8.1.0",
-    "eslint-plugin-prettier": "^3.3.1",
-    "jest": "^26.6.3",
-    "prettier": "^2.2.1",
+    "@nestjs/cli": "^9.0.0",
+    "@nestjs/schematics": "^9.0.0",
+    "@nestjs/testing": "^9.0.0",
+    "@types/bcrypt": "^5.0.0",
+    "@types/express": "^4.17.13",
+    "@types/jest": "29.2.4",
+    "@types/node": "18.11.18",
+    "@types/passport-jwt": "^3.0.8",
+    "@types/supertest": "^2.0.11",
+    "@typescript-eslint/eslint-plugin": "^5.0.0",
+    "@typescript-eslint/parser": "^5.0.0",
+    "eslint": "^8.0.1",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "jest": "29.3.1",
+    "prettier": "^2.3.2",
+    "source-map-support": "^0.5.20",
     "supertest": "^6.1.3",
-    "ts-jest": "^26.5.4",
-    "ts-loader": "^8.0.18",
-    "ts-node": "^9.1.1",
-    "tsconfig-paths": "^3.9.0",
-    "typescript": "^4.2.3"
+    "ts-jest": "29.0.3",
+    "ts-loader": "^9.2.3",
+    "ts-node": "^10.0.0",
+    "tsconfig-paths": "4.1.1",
+    "typescript": "^4.7.4"
   },
   "jest": {
     "moduleFileExtensions": [
