@@ -1,21 +1,22 @@
-import { TypeOrmModuleOptions } from "@nestjs/typeorm";
-import envConfig from "./env.config";
+import { DataSource, DataSourceOptions } from 'typeorm';
+require('dotenv').config()
 
-const seedConfig: TypeOrmModuleOptions = {
-  type: "postgres",
-  host: envConfig.DB_Host,
-  port: +envConfig.DB_Port,
-  username: envConfig.DB_UserName,
-  password: envConfig.DB_Password,
-  database: envConfig.DB_Name,
-  entities: [__dirname + "/../modules/**/entities/*.entity{.ts,.js}"],
-  migrations: [__dirname + "/../seed/*{.ts,.js}"],
-  cli: {
-    migrationsDir: "./src/seed",
-  },
+export const seedSourceOptions: DataSourceOptions = {
+  type: 'postgres',
+  host: process.env.DB_HOST,
+  port: +process.env.DB_PORT,
+  username: process.env.DB_USERNAME,
+  password: process.env.DB_PASSWORD,
+  database: process.env.DB_NAME,
+  entities: [__dirname + '/../modules/**/*.entity{.ts,.js}'],
+  migrations: [__dirname + '/../seed/*{.ts,.js}'],
   migrationsRun: false, // to automatically run migrations set it to true
   synchronize: false, //true if you want to create auto migrations
-  migrationsTableName: "seed",
-  keepConnectionAlive: true,
+  // logging: ["query", "error"],
+
+  migrationsTableName: 'seeds',
 };
-export default seedConfig;
+
+const seedSource = new DataSource(seedSourceOptions);
+
+export default seedSource;
