@@ -1,75 +1,87 @@
-import { Body, Controller, Get, Param, Post, Req, UseGuards, ValidationPipe } from "@nestjs/common";
-import { AuthGuard } from "@nestjs/passport";
-import { IsSuperAdmin } from "src/guards/isSuperAdmin.guard";
 import {
-  ChangePasswordDto,
-  ForgetPasswordDto,
-  LoginDto,
-  ResetPasswordDto,
-  UserDto,
-  ContactUsDto,
-} from "../users/dto/user.dto";
-import { AuthService } from "./auth.service";
-import { LoginAsDto } from "./dto/auth.dto";
+  Body,
+  Controller,
+  Post,
+  Param,
+  Req,
+  UseGuards,
+  UsePipes,
+  ValidationPipe,
+  Get,
+  Query,
+} from '@nestjs/common';
+import { JwtAuthGuard } from 'src/guards/jwt-auth.guard';
+import { CustomPipe } from 'src/pipe/customValidation.pipe';
+import { CreateUserDto } from '../users/dto/createUser.dto';
+import { AuthService } from './auth.service';
+import { ChangePasswordDto } from './dto/changePassword.dto';
+import { LoginDto } from './dto/loginUser.dto';
+import { ForgetPasswordDto } from './dto/forgetPassword.dto';
+import { ResetPasswordDto } from './dto/resetPassword.dto';
+import { RoleEnum } from 'src/common/enums/role.enum';
+import { Roles } from 'src/decorators/roles.decorator';
+import { RolesGuard } from 'src/guards/roles.guard';
 
-@Controller("auth")
+@Controller('auth')
 export class AuthController {
   constructor(private authService: AuthService) {}
 
-  @Post("/signup")
-  signUp(@Body(ValidationPipe) userDto: UserDto) {
+  @Post('/signup')
+  @UsePipes(CustomPipe)
+  signUp(@Body() userDto: CreateUserDto) {
     return this.authService.signUp(userDto);
   }
 
-  @Post("/signin")
+  @Post('/signin')
   signIn(@Body(ValidationPipe) userDto: LoginDto): Promise<object> {
     return this.authService.signIn(userDto);
   }
 
-  @UseGuards(AuthGuard())
-  @Get("/refresh-user")
-  refreshToken(@Req() req): Promise<object> {
-    return this.authService.refreshUser(req.user.email);
-  }
-
-  @Post("/change-password")
-  @UseGuards(AuthGuard())
-  async changePassword(@Body(ValidationPipe) userDto: ChangePasswordDto, @Req() req): Promise<{ message: string }> {
-    const status = await this.authService.changePassword(userDto.oldPassword, userDto.newPassword, req.user);
+  @Post('/change-password')
+  @UseGuards(JwtAuthGuard)
+  async changePassword(
+    @Body(ValidationPipe) userDto: ChangePasswordDto,
+    @Req() req,
+  ): Promise<{ message: string }> {
+    const status = await this.authService.changePassword(
+      userDto.oldPassword,
+      userDto.newPassword,
+      req.user,
+    );
     return { message: status };
   }
 
-  @Post("/test")
-  @UseGuards(AuthGuard())
-  test(@Req() req) {
-    return req.user;
-  }
-
-  @Post("/forget-password")
-  async forgetPassword(@Body(ValidationPipe) forgetPasswordDto: ForgetPasswordDto): Promise<{ message: any }> {
-    const status = await this.authService.forgetPassword(forgetPasswordDto.email);
+  @Post('/forget-password')
+  @UseGuards(JwtAuthGuard)
+  async forgetPassword(
+    @Body(ValidationPipe) forgetPasswordDto: ForgetPasswordDto,
+  ): Promise<{ message: any }> {
+    const status = await this.authService.forgetPassword(
+      forgetPasswordDto.email,
+    );
     return { message: status };
   }
 
-  @Post("/reset-password/:token")
+  @Post('/reset-password/:token')
+  @UseGuards(JwtAuthGuard)
   async resetPassword(
     @Body(ValidationPipe) resetPasswordDto: ResetPasswordDto,
-    @Param() param
+    @Param() param,
   ): Promise<{ message: any }> {
-    const status = await this.authService.resetPassword(resetPasswordDto.password, param.token);
-    return { message: status };
-  }
-  @Post("/contact-us")
-  async contactUs(@Body(ValidationPipe) contactUsDto: ContactUsDto): Promise<{ message: any }> {
-    const status = await this.authService.contactUs(contactUsDto);
+    const status = await this.authService.resetPassword(
+      resetPasswordDto.password,
+      param.token,
+    );
+
     return { message: status };
   }
 
-  // @UseGuards(IsSuperAdmin)
-  // @UseGuards(AuthGuard())
-  // @Post("/login-as/:orgId")
-  // loginAs(@Param(ValidationPipe) param: LoginAsDto) {
-  //   const { orgId } = param;
-  //   return this.authService.loginAs(orgId);
-  // }
+  @Get('verify-token')
+  @UseGuards(JwtAuthGuard)
+  async verifyToken(
+    @Query('token') encryptedEmail: string,
+  ): Promise<{ email: string }> {
+    const decryptedEmail = await this.authService.decryptToken(encryptedEmail);
+    return { email: decryptedEmail };
+  }
 }
