@@ -1,26 +1,26 @@
-import { Module } from "@nestjs/common";
-import { JwtModule } from "@nestjs/jwt";
-import { PassportModule } from "@nestjs/passport";
-import { TypeOrmModule } from "@nestjs/typeorm";
+import { Module } from '@nestjs/common';
+import { AuthService } from './auth.service';
+import { AuthController } from './auth.controller';
+
+import { UsersModule } from '../users/users.module';
 import { MailModule } from "../mail/mail.module";
-import { UserRepository } from "../users/user.repository";
-import { AuthController } from "./auth.controller";
-import { AuthService } from "./auth.service";
-import { jwtConstants } from "./constants";
-import { JwtStrategy } from "./jwt.strategy";
+import { JwtModule } from '@nestjs/jwt';
+
+import { PassportModule } from '@nestjs/passport';
+import { JwtStrategy } from './jwt.strategy';
+import { jwtConstants } from 'src/generalUtils/constant';
 
 @Module({
   imports: [
-    PassportModule.register({ defaultStrategy: "jwt" }),
+    UsersModule,
+    MailModule,
+    PassportModule,
     JwtModule.register({
       secret: jwtConstants.secret,
-      signOptions: { expiresIn: "200h" }, // 3600 },
+      signOptions: { expiresIn: jwtConstants.time },
     }),
-    TypeOrmModule.forFeature([UserRepository]),
-    MailModule,
   ],
-  controllers: [AuthController],
   providers: [AuthService, JwtStrategy],
-  exports: [JwtStrategy, PassportModule],
+  controllers: [AuthController],
 })
 export class AuthModule {}
