@@ -1,115 +1,111 @@
-import { HttpException, HttpStatus, Injectable, NotFoundException, UnauthorizedException } from "@nestjs/common";
-import { JwtService } from "@nestjs/jwt";
-import { InjectRepository } from "@nestjs/typeorm";
-import * as moment from "moment";
-import { v4 } from "uuid";
-import { MailService } from "../mail/mail.service";
-import { ContactUsDto, LoginDto, UserDto } from "../users/dto/user.dto";
-import { UserRepository } from "../users/user.repository";
-import { jwtPayload } from "./jwt-payload.interface";
+import {
+  HttpException,
+  HttpStatus,
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
+import { JwtService } from '@nestjs/jwt';
+
+import { CreateUserDto } from '../users/dto/createUser.dto';
+
+import { UsersService } from '../users/users.service';
+import { MailService } from '../mail/mail.service';
+import { LoginDto } from './dto/loginUser.dto';
+import { jwtPayload } from './jwt-payload.interface';
+import * as bcrypt from 'bcrypt';
+import { User } from '../users/entities/user.entity';
 
 @Injectable()
 export class AuthService {
   constructor(
-    @InjectRepository(UserRepository)
-    private userRepository: UserRepository,
+    private userService: UsersService,
+    private mailService: MailService,
     private jwtService: JwtService,
-    private mailService: MailService
   ) {}
 
-  async signUp(userDto: UserDto): Promise<object> {
-    const name = userDto.firstName + " " + userDto.lastName;
-    const res = await this.userRepository.signUp(userDto);
-    if (res) {
-      const email = res.email;
+  async signUp(userDto: CreateUserDto): Promise<object> {
+    const result = await this.userService.signUp(userDto);
+
+    if (result) {
+      const email = result.email;
       const payload: jwtPayload = { email };
       const accessToken = await this.jwtService.sign(payload);
-      delete res.password;
-      delete res.salt;
+      delete result.password;
       const response = {
-        ...res,
+        user: result,
         apiToken: accessToken,
-        name: res.lastName ? res.firstName + " " + res.lastName : res.firstName,
       };
       return response;
     } else {
-      throw new HttpException("User Not Created", HttpStatus.BAD_REQUEST);
+      throw new HttpException('User Not Created', HttpStatus.BAD_REQUEST);
     }
   }
   async signIn(userDto: LoginDto): Promise<object> {
-    let user = await this.userRepository.validateUserPassword(userDto);
+    let user = await this.userService.validateUser(userDto);
+
     if (!user) {
-      throw new UnauthorizedException("Invalid Credentials");
+      throw new UnauthorizedException('Invalid Credentials');
     }
-    const email = user.email;
-    const payload: jwtPayload = { email };
+    const payload: jwtPayload = { email: user.email };
     const accessToken = await this.jwtService.sign(payload);
-    delete user.password;
-    delete user.salt;
+    const { password, ...userWithoutPassword } = user;
 
     const response = {
-      ...user,
+      user: {...userWithoutPassword,
+        name: user.lastName
+        ? user.firstName + ' ' + user.lastName
+        : user.firstName
+      },
       apiToken: accessToken,
-      name: user.lastName ? user.firstName + " " + user.lastName : user.firstName,
     };
 
     return response;
   }
 
-  async refreshUser(userEmail): Promise<object> {
-    const user = await this.userRepository.findOne(
-      { email: userEmail },
-      {
-        select: [
-          "id",
-          "email",
-          "password",
-          "salt",
-          "firstName",
-          "lastName",
-          "role",
-          "subscriptionStatus",
-          "uuid",
-          "planId",
-        ],
-      }
-    );
-    if (!user) throw new NotFoundException("Email doest not Exists");
+  async changePassword(
+    oldPassword: string,
+    newPassword: string,
+    user: User,
+  ): Promise<string> {
+    if (oldPassword === newPassword)
+      throw new HttpException(
+        'Old Password cannot be new password',
+        HttpStatus.BAD_REQUEST,
+      );
 
-    const email = user.email;
-    const payload: jwtPayload = { email };
-    const accessToken = await this.jwtService.sign(payload);
-    delete user.password;
-    delete user.salt;
-    const response = {
-      ...user,
-      apiToken: accessToken,
-      name: user.lastName ? user.firstName + " " + user.lastName : user.firstName,
-    };
+    if (user && bcrypt.compareSync(oldPassword, user.password)) {
+      newPassword = bcrypt.hashSync(newPassword, bcrypt.genSaltSync());
 
-    return response;
-  }
-  async changePassword(oldPassword: string, newPassword: string, user: any): Promise<string> {
-    return await this.userRepository.changePasswordUser(oldPassword, newPassword, user);
+      user.password = newPassword;
+      await this.userService.saveUser(user);
+      return 'Succesfully Updated';
+    } else {
+      throw new HttpException(
+        'Invalid User or invalid old password',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
   }
 
   async forgetPassword(email) {
-    const user = await this.userRepository.findOne(
-      { email },
-      { select: ["id", "email", "password", "salt", "firstName", "lastName", "role", "forgetPasswordToken", "expires"] }
+    const user = await this.userService.forgetPasswordUser(email);
+    if (!user) throw new NotFoundException('Email doest not Exists');
+    return await this.mailService.sendForgetPassword(
+      user,
+      user.forgetPasswordToken,
     );
-    if (!user) throw new NotFoundException("Email doest not Exists");
-    //creating new token
-    user.forgetPasswordToken = v4();
-    user.expires = moment().add(10, "minutes").unix();
-    await user.save();
-    return await this.mailService.sendForgetPassword(user, user.forgetPasswordToken);
-  }
-  async contactUs(contactUsDto: ContactUsDto) {
-    return await this.mailService.sendContactUs(contactUsDto);
   }
 
   async resetPassword(password, token) {
-    return await this.userRepository.resetPassword(token, password);
+    return await this.userService.resetPassword(token, password);
+  }
+
+  async decryptToken(token) {
+    try {
+      return this.jwtService.verify(token).email;
+    } catch (error) {
+      throw new UnauthorizedException('Invalid token');
+    }
   }
 }
